move funcs from test/e2e/framework/job to test/e2e/apps
