+++ test/e2e/framework/job/rest.go
@@ -17,15 +17,11 @@ limitations under the License.
 package job
 
 import (
-	"fmt"
-
 	batchv1 "k8s.io/api/batch/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/util/wait"
 	clientset "k8s.io/client-go/kubernetes"
-	"k8s.io/kubernetes/test/e2e/framework"
 )
 
 // GetJob uses c to get the Job in namespace ns named name. If the returned error is nil, the returned Job is valid.
@@ -50,29 +46,6 @@ func CreateJob(c clientset.Interface, ns string, job *batchv1.Job) (*batchv1.Job
 // been updated.
 func UpdateJob(c clientset.Interface, ns string, job *batchv1.Job) (*batchv1.Job, error) {
 	return c.BatchV1().Jobs(ns).Update(job)
-}
-
-// UpdateJobWithRetries updates job with retries.
-func UpdateJobWithRetries(c clientset.Interface, namespace, name string, applyUpdate func(*batchv1.Job)) (job *batchv1.Job, err error) {
-	jobs := c.BatchV1().Jobs(namespace)
-	var updateErr error
-	pollErr := wait.PollImmediate(framework.Poll, JobTimeout, func() (bool, error) {
-		if job, err = jobs.Get(name, metav1.GetOptions{}); err != nil {
-			return false, err
-		}
-		// Apply the update, then attempt to push it to the apiserver.
-		applyUpdate(job)
-		if job, err = jobs.Update(job); err == nil {
-			framework.Logf("Updating job %s", name)
-			return true, nil
-		}
-		updateErr = err
-		return false, nil
-	})
-	if pollErr == wait.ErrWaitTimeout {
-		pollErr = fmt.Errorf("couldn't apply the provided updated to job %q: %v", name, updateErr)
-	}
-	return job, pollErr
 }
 
 // DeleteJob uses c to delete the Job named name in namespace ns. If the returned error is nil, the Job has been
