+++ test/e2e/node/ttlafterfinished.go
@@ -17,11 +17,14 @@ limitations under the License.
 package node
 
 import (
+	"fmt"
 	"time"
 
 	batchv1 "k8s.io/api/batch/v1"
 	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/wait"
+	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/kubernetes/pkg/util/slice"
 	"k8s.io/kubernetes/test/e2e/framework"
 	jobutil "k8s.io/kubernetes/test/e2e/framework/job"
@@ -29,7 +32,12 @@ import (
 	"github.com/onsi/ginkgo"
 )
 
-const dummyFinalizer = "k8s.io/dummy-finalizer"
+const (
+	dummyFinalizer = "k8s.io/dummy-finalizer"
+
+	// JobTimeout is how long to wait for a job to finish.
+	JobTimeout = 15 * time.Minute
+)
 
 var _ = framework.KubeDescribe("[Feature:TTLAfterFinished][NodeAlphaFeature:TTLAfterFinished]", func() {
 	f := framework.NewDefaultFramework("ttlafterfinished")
@@ -47,7 +55,7 @@ func cleanupJob(f *framework.Framework, job *batchv1.Job) {
 	removeFinalizerFunc := func(j *batchv1.Job) {
 		j.ObjectMeta.Finalizers = slice.RemoveString(j.ObjectMeta.Finalizers, dummyFinalizer, nil)
 	}
-	_, err := jobutil.UpdateJobWithRetries(c, ns, job.Name, removeFinalizerFunc)
+	_, err := updateJobWithRetries(c, ns, job.Name, removeFinalizerFunc)
 	framework.ExpectNoError(err)
 	jobutil.WaitForJobGone(c, ns, job.Name, wait.ForeverTestTimeout)
 
@@ -78,13 +86,13 @@ func testFinishedJob(f *framework.Framework) {
 	framework.ExpectNoError(err)
 
 	framework.Logf("Wait for TTL after finished controller to delete the Job")
-	err = jobutil.WaitForJobDeleting(c, ns, job.Name)
+	err = waitForJobDeleting(c, ns, job.Name)
 	framework.ExpectNoError(err)
 
 	framework.Logf("Check Job's deletionTimestamp and compare with the time when the Job finished")
 	job, err = jobutil.GetJob(c, ns, job.Name)
 	framework.ExpectNoError(err)
-	finishTime := jobutil.FinishTime(job)
+	finishTime := FinishTime(job)
 	finishTimeUTC := finishTime.UTC()
 	framework.ExpectNotEqual(finishTime.IsZero(), true)
 
@@ -93,4 +101,50 @@ func testFinishedJob(f *framework.Framework) {
 
 	expireAtUTC := finishTimeUTC.Add(time.Duration(ttl) * time.Second)
 	framework.ExpectEqual(deleteAtUTC.Before(expireAtUTC), false)
+}
+
+// FinishTime returns finish time of the specified job.
+func FinishTime(finishedJob *batchv1.Job) metav1.Time {
+	var finishTime metav1.Time
+	for _, c := range finishedJob.Status.Conditions {
+		if (c.Type == batchv1.JobComplete || c.Type == batchv1.JobFailed) && c.Status == v1.ConditionTrue {
+			return c.LastTransitionTime
+		}
+	}
+	return finishTime
+}
+
+// updateJobWithRetries updates job with retries.
+func updateJobWithRetries(c clientset.Interface, namespace, name string, applyUpdate func(*batchv1.Job)) (job *batchv1.Job, err error) {
+	jobs := c.BatchV1().Jobs(namespace)
+	var updateErr error
+	pollErr := wait.PollImmediate(framework.Poll, JobTimeout, func() (bool, error) {
+		if job, err = jobs.Get(name, metav1.GetOptions{}); err != nil {
+			return false, err
+		}
+		// Apply the update, then attempt to push it to the apiserver.
+		applyUpdate(job)
+		if job, err = jobs.Update(job); err == nil {
+			framework.Logf("Updating job %s", name)
+			return true, nil
+		}
+		updateErr = err
+		return false, nil
+	})
+	if pollErr == wait.ErrWaitTimeout {
+		pollErr = fmt.Errorf("couldn't apply the provided updated to job %q: %v", name, updateErr)
+	}
+	return job, pollErr
+}
+
+// waitForJobDeleting uses c to wait for the Job jobName in namespace ns to have
+// a non-nil deletionTimestamp (i.e. being deleted).
+func waitForJobDeleting(c clientset.Interface, ns, jobName string) error {
+	return wait.PollImmediate(framework.Poll, JobTimeout, func() (bool, error) {
+		curr, err := c.BatchV1().Jobs(ns).Get(jobName, metav1.GetOptions{})
+		if err != nil {
+			return false, err
+		}
+		return curr.ObjectMeta.DeletionTimestamp != nil, nil
+	})
 }
